refactor userDetail and userDetailService
