+++ src/main/java/com/skillstorm/linkedinclone/models/CustomUserDetails.java
@@ -1,0 +1,42 @@
+package com.skillstorm.linkedinclone.models;
+
+import lombok.Data;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.Collection;
+
+
+@Data
+public class CustomUserDetails implements UserDetails {
+    private int id;
+    private String email;
+    private String password;
+    private Collection<? extends GrantedAuthority> authorities;
+
+
+    @Override
+    public String getUsername() {
+        return email;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
