+++ src/main/java/com/skillstorm/linkedinclone/security/JWTAuthenticationFilter.java
@@ -1,5 +1,6 @@
 package com.skillstorm.linkedinclone.security;
 
+import com.skillstorm.linkedinclone.services.CustomUserDetailsService;
 import com.skillstorm.linkedinclone.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -20,7 +21,7 @@ public class JWTAuthenticationFilter extends OncePerRequestFilter {
     @Autowired
     private JWTGenerator tokenGenerator;
     @Autowired
-    private UserService userService;
+    private CustomUserDetailsService userService;
 
     @Override
     protected void doFilterInternal(HttpServletRequest request,
