+++ src/main/java/com/skillstorm/linkedinclone/services/CustomUserDetailsService.java
@@ -1,0 +1,38 @@
+package com.skillstorm.linkedinclone.services;
+
+import com.skillstorm.linkedinclone.models.CustomUserDetails;
+import com.skillstorm.linkedinclone.models.User;
+import com.skillstorm.linkedinclone.repositories.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.Collections;
+import java.util.List;
+
+@Service
+public class CustomUserDetailsService implements UserDetailsService {
+
+    @Autowired
+    private UserRepository userRepository;
+
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        User user = userRepository.findByEmail(username).orElseThrow(()-> new UsernameNotFoundException(username+ " not found!"));
+        List<SimpleGrantedAuthority> authorities = Collections.singletonList(new SimpleGrantedAuthority(user.getRole()));
+        return mapUserToCustomUserDetails(user, authorities);
+    }
+
+    private CustomUserDetails mapUserToCustomUserDetails(User user, List<SimpleGrantedAuthority> authorities){
+        CustomUserDetails customUserDetails = new CustomUserDetails();
+        customUserDetails.setId(user.getId());
+        customUserDetails.setEmail(user.getEmail());
+        customUserDetails.setPassword(user.getPassword());
+        customUserDetails.setAuthorities(authorities);
+        return customUserDetails;
+    }
+}
